student-progress-tracker-backend/requirements.txt
FastAPI
SQLAlchemy
pydantic
python-dotenv
pymysql
bcrypt==4.3.0
python-jose
uvicorn
pandas
openpyxl
websockets
alembic
pytest
httpx
beautifulsoup4
pymysql
email-validator
passlib
python-multipart
mailjet-rest
cryptography
pypdf
pdfplumber